[DSL] Fixed the incorrect additiveness of the `bool` filter and query
Due to the implementation of the `must`, `should` and `must_not` methods
of the {Bool} query and filter, repeated invocation of the `to_hash` method
changed the inner Hash representation of the `@value` in an incorrect way.

The support for multiple invocations of methods within the `must`, `should` and `must_not`
methods has been removed, documentation ands tests has been changed accordingly.